+++ system/service/apigw_route.go
@@ -2,15 +2,9 @@ package service
 
 import (
 	"context"
-	"encoding/base64"
-	"errors"
-	"io/ioutil"
-	"math"
-	"time"
 
 	"github.com/cortezaproject/corteza-server/pkg/actionlog"
 	"github.com/cortezaproject/corteza-server/pkg/apigw"
-	"github.com/cortezaproject/corteza-server/pkg/apigw/profiler"
 	a "github.com/cortezaproject/corteza-server/pkg/auth"
 
 	"github.com/cortezaproject/corteza-server/store"
@@ -258,161 +252,4 @@ func (svc *apigwRoute) Search(ctx context.Context, filter types.ApigwRouteFilter
 	}()
 
 	return r, f, svc.recordAction(ctx, aProps, ApigwRouteActionSearch, err)
-}
-
-// HitsAggregated fetches a list of hits from integration gateway profiler
-func (svc *apigwRoute) Hits(ctx context.Context, filter types.ApigwProfilerFilter) (r types.ApigwProfilerHitSet, f types.ApigwProfilerFilter, err error) {
-
-	f = filter
-	r = make(types.ApigwProfilerHitSet, 0)
-
-	uDec, err := base64.URLEncoding.DecodeString(filter.Path)
-
-	if err != nil {
-		return
-	}
-
-	filter.Path = string(uDec)
-
-	if filter.Path == "" && filter.Hit == "" {
-		err = errors.New("fetching all hits (no route and hit specified) not supported")
-		return
-	}
-
-	var sorting = profiler.Sort{
-		Hit:    filter.Hit,
-		Path:   filter.Path,
-		Before: filter.Before,
-	}
-
-	var (
-		list = apigw.Service().Profiler().Hits(sorting)
-	)
-
-	var pp = ""
-
-	for k, _ := range list {
-		if filter.Hit != "" {
-			pp = k
-			break
-		}
-
-		if filter.Path != "" && k == filter.Path {
-			pp = k
-			break
-		}
-	}
-
-	if pp == "" {
-		return
-	}
-
-	for _, h := range list[pp] {
-		hh := &types.ApigwProfilerHit{
-			ID: h.ID,
-
-			Route:   h.Route,
-			Status:  h.Status,
-			Request: *h.R,
-
-			Ts: h.Ts,
-			Tf: h.Tf,
-			D:  h.D,
-			Dr: float64(h.D.Microseconds()) / 1000,
-		}
-
-		// fetch body only on hit details
-		if filter.Hit != "" {
-			hh.Body, _ = ioutil.ReadAll(hh.Request.Body)
-		}
-
-		r = append(r, hh)
-	}
-
-	// sort
-	profiler.SortHits(&r, &f)
-
-	// filter sorted
-	profiler.FilterHits(&r, &f)
-
-	return
-}
-
-// HitsAggregated fetches a list of hits from integration gateway profiler
-// and aggregates them with assigned filters
-func (svc *apigwRoute) HitsAggregated(ctx context.Context, filter types.ApigwProfilerFilter) (r types.ApigwProfilerAggregationSet, f types.ApigwProfilerFilter, err error) {
-	f = filter
-	r = make(types.ApigwProfilerAggregationSet, 0)
-
-	uDec, err := base64.URLEncoding.DecodeString(filter.Path)
-
-	if err != nil {
-		return
-	}
-
-	filter.Path = string(uDec)
-
-	var (
-		list = apigw.Service().Profiler().Hits(profiler.Sort{
-			Path:   filter.Path,
-			Before: filter.Before,
-		})
-
-		tsum, tmin, tmax time.Duration
-		ssum, smin, smax int64
-		i                uint64 = 1
-	)
-
-	for p, v := range list {
-		tmin, tmax, tsum = time.Hour, 0, 0
-		smin, smax, ssum = math.MaxInt64, 0, 0
-
-		i = 0
-
-		for _, vv := range v {
-			var (
-				d = *vv.D
-				s = vv.R.ContentLength
-			)
-
-			if d < tmin {
-				tmin = d
-			}
-
-			if d > tmax {
-				tmax = d
-			}
-
-			if s < smin {
-				smin = s
-			}
-
-			if s > smax {
-				smax = s
-			}
-
-			tsum += d
-			ssum += s
-			i++
-		}
-
-		r = append(r, &types.ApigwProfilerAggregation{
-			Path:  p,
-			Count: i,
-			Tmin:  float64(tmin.Microseconds()) / 1000,
-			Tmax:  float64(tmax.Microseconds()) / 1000,
-			Tavg:  float64(tsum.Microseconds()) / float64(i) / 1000,
-			Smin:  smin,
-			Smax:  smax,
-			Savg:  float64(ssum) / float64(i),
-		})
-	}
-
-	// sort
-	profiler.SortAggregation(&r, &f)
-
-	// filter
-	profiler.FilterAggregation(&r, &f)
-
-	return
 }
